added createmodulecontroller and partner zone controller
