+++ src/Controller/PortalController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\Language;
 use App\Entity\LearningModule;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
 class PortalController extends AbstractController
@@ -12,7 +13,7 @@ class PortalController extends AbstractController
     /**
      * @Route("/portal", name="portal")
      */
-    public function index()
+    public function index(): Response
     {
         $language = $this->getDoctrine()->getRepository(Language::class)->find(1);
         $modules = $this->getDoctrine()->getRepository(LearningModule::class)->findBy(['isPublished' => true]);
