+++ templates/create_module/index.html.twig
@@ -1,0 +1,26 @@
+{% extends 'base.html.twig' %}
+
+{% block title %}Hello CreateModuleController!{% endblock %}
+
+{% block body %}
+<style>
+    .example-wrapper { margin: 1em auto; max-width: 800px; width: 95%; font: 18px/1.5 sans-serif; }
+    .example-wrapper code { background: #F5F5F5; padding: 2px 6px; }
+</style>
+
+<div class="example-wrapper">
+    <h1>Howdy Partner!</h1>
+    <h2>To create a new module, please fill in one or more of the translations</h2>
+    {{ form_start(form) }}
+    {% for translation in form.translations %}
+        <hr>
+        <h2>{{translation.vars.value.language.name}}</h2>
+        {{ form_row(translation) }}
+    {% endfor %}
+    <hr>
+    {#        {{ form_row(form.image) }}#}
+    {{ form_row(form.badge) }}
+    {{ form_row(form.create) }}
+    {{ form_end(form) }}
+</div>
+{% endblock %}
